squash! Make Ensemble n_neurons/dimensions Unconfigurable
NdarrayParam should not need to know about the specific attribute names
that might be used in the shape specification. All we need to do is
check for strings as any string will refer to an attribute and thus
cannot be coerced/validated on a Config. This makes this check also
more general in case such a parameter is used for other classes.